+++ CoreHook.sln
@@ -57,6 +57,10 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "test", "test", "{651578A4-4
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CoreHook.Tests", "test\CoreHook.Tests\CoreHook.Tests.csproj", "{AAAC3722-8911-4F80-8232-840D81F8C938}"
 EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "hook", "hook", "{BFEBEA69-1024-42F8-A2D5-B1B43E6CE719}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CoreHook.Tests.SimpleHook1", "test\hook\CoreHook.Tests.SimpleHook1\CoreHook.Tests.SimpleHook1.csproj", "{6210E5FC-5876-4A0C-AACE-12C1BB63CF5A}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -75,8 +79,8 @@ Global
 		{F5C8C0B0-3DC0-40A8-B2B0-3E8A352092AF}.Release|x86.Build.0 = Release|x86
 		{8A1CF652-4A6B-4CCA-878E-0F1B6A549EC1}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{8A1CF652-4A6B-4CCA-878E-0F1B6A549EC1}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{8A1CF652-4A6B-4CCA-878E-0F1B6A549EC1}.Debug|x86.ActiveCfg = Debug|x86
-		{8A1CF652-4A6B-4CCA-878E-0F1B6A549EC1}.Debug|x86.Build.0 = Debug|x86
+		{8A1CF652-4A6B-4CCA-878E-0F1B6A549EC1}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{8A1CF652-4A6B-4CCA-878E-0F1B6A549EC1}.Debug|x86.Build.0 = Debug|Any CPU
 		{8A1CF652-4A6B-4CCA-878E-0F1B6A549EC1}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{8A1CF652-4A6B-4CCA-878E-0F1B6A549EC1}.Release|Any CPU.Build.0 = Release|Any CPU
 		{8A1CF652-4A6B-4CCA-878E-0F1B6A549EC1}.Release|x86.ActiveCfg = Release|x86
@@ -135,8 +139,8 @@ Global
 		{5CD24098-AAA6-4273-8714-FA61E4E1F0D1}.Release|x86.Build.0 = Release|Any CPU
 		{AE8E24A5-72F5-4FC7-9095-49700AF38570}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{AE8E24A5-72F5-4FC7-9095-49700AF38570}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{AE8E24A5-72F5-4FC7-9095-49700AF38570}.Debug|x86.ActiveCfg = Debug|x86
-		{AE8E24A5-72F5-4FC7-9095-49700AF38570}.Debug|x86.Build.0 = Debug|x86
+		{AE8E24A5-72F5-4FC7-9095-49700AF38570}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{AE8E24A5-72F5-4FC7-9095-49700AF38570}.Debug|x86.Build.0 = Debug|Any CPU
 		{AE8E24A5-72F5-4FC7-9095-49700AF38570}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{AE8E24A5-72F5-4FC7-9095-49700AF38570}.Release|Any CPU.Build.0 = Release|Any CPU
 		{AE8E24A5-72F5-4FC7-9095-49700AF38570}.Release|x86.ActiveCfg = Release|x86
@@ -151,8 +155,8 @@ Global
 		{F6C235DE-F333-4FAC-824B-8310AB442066}.Release|x86.Build.0 = Release|Any CPU
 		{63EE7CAD-89B3-49AA-81E4-FECFFA9F7B8B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{63EE7CAD-89B3-49AA-81E4-FECFFA9F7B8B}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{63EE7CAD-89B3-49AA-81E4-FECFFA9F7B8B}.Debug|x86.ActiveCfg = Debug|x86
-		{63EE7CAD-89B3-49AA-81E4-FECFFA9F7B8B}.Debug|x86.Build.0 = Debug|x86
+		{63EE7CAD-89B3-49AA-81E4-FECFFA9F7B8B}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{63EE7CAD-89B3-49AA-81E4-FECFFA9F7B8B}.Debug|x86.Build.0 = Debug|Any CPU
 		{63EE7CAD-89B3-49AA-81E4-FECFFA9F7B8B}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{63EE7CAD-89B3-49AA-81E4-FECFFA9F7B8B}.Release|Any CPU.Build.0 = Release|Any CPU
 		{63EE7CAD-89B3-49AA-81E4-FECFFA9F7B8B}.Release|x86.ActiveCfg = Release|x86
@@ -167,16 +171,16 @@ Global
 		{83EEB4CF-E21D-4187-9D73-E9A9891268DE}.Release|x86.Build.0 = Release|Any CPU
 		{B6EA0C64-B13B-4F14-B510-DFB616F5225C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{B6EA0C64-B13B-4F14-B510-DFB616F5225C}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{B6EA0C64-B13B-4F14-B510-DFB616F5225C}.Debug|x86.ActiveCfg = Debug|x86
-		{B6EA0C64-B13B-4F14-B510-DFB616F5225C}.Debug|x86.Build.0 = Debug|x86
+		{B6EA0C64-B13B-4F14-B510-DFB616F5225C}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{B6EA0C64-B13B-4F14-B510-DFB616F5225C}.Debug|x86.Build.0 = Debug|Any CPU
 		{B6EA0C64-B13B-4F14-B510-DFB616F5225C}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{B6EA0C64-B13B-4F14-B510-DFB616F5225C}.Release|Any CPU.Build.0 = Release|Any CPU
 		{B6EA0C64-B13B-4F14-B510-DFB616F5225C}.Release|x86.ActiveCfg = Release|x86
 		{B6EA0C64-B13B-4F14-B510-DFB616F5225C}.Release|x86.Build.0 = Release|x86
 		{C1446EE3-AF70-4902-B2CE-ED7F53FC0CB9}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{C1446EE3-AF70-4902-B2CE-ED7F53FC0CB9}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{C1446EE3-AF70-4902-B2CE-ED7F53FC0CB9}.Debug|x86.ActiveCfg = Debug|x86
-		{C1446EE3-AF70-4902-B2CE-ED7F53FC0CB9}.Debug|x86.Build.0 = Debug|x86
+		{C1446EE3-AF70-4902-B2CE-ED7F53FC0CB9}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{C1446EE3-AF70-4902-B2CE-ED7F53FC0CB9}.Debug|x86.Build.0 = Debug|Any CPU
 		{C1446EE3-AF70-4902-B2CE-ED7F53FC0CB9}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{C1446EE3-AF70-4902-B2CE-ED7F53FC0CB9}.Release|Any CPU.Build.0 = Release|Any CPU
 		{C1446EE3-AF70-4902-B2CE-ED7F53FC0CB9}.Release|x86.ActiveCfg = Release|x86
@@ -197,6 +201,14 @@ Global
 		{AAAC3722-8911-4F80-8232-840D81F8C938}.Release|Any CPU.Build.0 = Release|Any CPU
 		{AAAC3722-8911-4F80-8232-840D81F8C938}.Release|x86.ActiveCfg = Release|x86
 		{AAAC3722-8911-4F80-8232-840D81F8C938}.Release|x86.Build.0 = Release|x86
+		{6210E5FC-5876-4A0C-AACE-12C1BB63CF5A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{6210E5FC-5876-4A0C-AACE-12C1BB63CF5A}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{6210E5FC-5876-4A0C-AACE-12C1BB63CF5A}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{6210E5FC-5876-4A0C-AACE-12C1BB63CF5A}.Debug|x86.Build.0 = Debug|Any CPU
+		{6210E5FC-5876-4A0C-AACE-12C1BB63CF5A}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{6210E5FC-5876-4A0C-AACE-12C1BB63CF5A}.Release|Any CPU.Build.0 = Release|Any CPU
+		{6210E5FC-5876-4A0C-AACE-12C1BB63CF5A}.Release|x86.ActiveCfg = Release|Any CPU
+		{6210E5FC-5876-4A0C-AACE-12C1BB63CF5A}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -222,6 +234,8 @@ Global
 		{C1446EE3-AF70-4902-B2CE-ED7F53FC0CB9} = {397BC2E9-F70B-4C68-A6B3-405AFB6123CE}
 		{9BF07061-551A-431F-B27B-AD35E7AA7DA8} = {397BC2E9-F70B-4C68-A6B3-405AFB6123CE}
 		{AAAC3722-8911-4F80-8232-840D81F8C938} = {651578A4-4B76-4A11-897F-45A3D5714574}
+		{BFEBEA69-1024-42F8-A2D5-B1B43E6CE719} = {651578A4-4B76-4A11-897F-45A3D5714574}
+		{6210E5FC-5876-4A0C-AACE-12C1BB63CF5A} = {BFEBEA69-1024-42F8-A2D5-B1B43E6CE719}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {7F1E6A8F-CBCB-4F01-B2C4-446C6F3AD674}
