+++ test/CoreHook.Tests/NamedPipeTest.cs
@@ -5,7 +5,7 @@ using Xunit;
 
 namespace CoreHook.Tests
 {
-    public class RemoteHookingTests
+    public class NamedPipeTest
     {
     }
 }
